Names map uses dynamic clustering; fix focus-ring clipping (#4159)
- Extract `Observations::MapsController#index`'s clustered-map loader
  + JSON refetch payload into a `ClusteredObservationMap` concern.
- `Names::MapsController#show` now includes the concern, runs the
  same cap+banner+refetch pipeline, and responds to `format.json` so
  the client can refetch on viewport change. The ERB view drops the
  server-side bucketing approach in favor of `clustering: true`,
  matching `observations/maps/index`. `/names/:id/map` now uses the
  same pipeline as the obs-index map. Other obs-mapping views
  audited; only this one was still on server-side bucketing.
- Focus-ring clipping fixes:
  - `.gm-style-iw-d` content has `overflow: auto`; focusables inside
    (links, buttons in obs / cluster popups) now draw their outline
    inside the box via `outline-offset: -2px` so the popup container
    can't clip it.
  - `.map-popup-cluster-name` was using `overflow: hidden; ellipsis`
    which clipped the right side of the link's focus ring. Switched
    to `overflow-wrap: break-word` so long names wrap rather than
    truncate.
  - `.sidebar-nav .list-group-item` and `#logo_link` get
    `outline-offset: -2px` on `:focus-visible` so the keyboard focus
    ring isn't clipped by the existing overflow constraints.
